moved get_group to common_model since both group_model and home_model share it
