+++ app/hindsite/home/home_model.py
@@ -59,23 +59,6 @@ def accept_invitation(membership: Membership):
     db.session.commit()
 
 
-def send_invitation(group_id: int, email: str):
-    """
-    Creates a Membership that signals an invitation to a user, by default the membership
-    is not an ownership membership and will have <code>invitation_accepted</code> set to False
-
-    :param group_id: The id of the group attached to the membership.
-    :param email: The email of the user to be added to the membership.
-    :returns: **Membership** A reference to the membership object.
-    """
-    user = get_user(email)
-    group = get_group(group_id)
-    membership = Membership(user, group)
-    db.session.add(membership)
-    db.session.commit()
-    return membership
-
-
 def get_group(group_id: int):
     """
     Gets a group record belonging to a user.
