+++ app/hindsite/home/templates/home.html
@@ -38,6 +38,10 @@
                 </span>
             </li>
         </ul>
+        <div id="invites"
+             hx-get="/invites">
+
+        </div>
         <div 
             id="modal"
             class="modal modal-blur fade"
