+++ app/hindsite/home/templates/partials/invites.html
@@ -1,0 +1,3 @@
+{% for invite in invites %}
+    <p>{{ invite }}</p>
+{% endfor %}
